+++ CHANGELOG.md
@@ -1,6 +1,10 @@
  
 # QOwnNotes Changelog
 
+## 0.37
+- trashed notes can now be viewed and restored from your ownCloud server
+	- please install QOwnNotesAPI v0.3 for trash support
+
 ## 0.36
 - fixed a problem with the OSX build